size_t -> unsigned long int
git-svn-id: https://psvn.ims.uni-hannover.de/svn/fucap/trunk@141 1233f08c-2788-c397-e086-bd64547a093a

+++ src/l_ftos.cc
@@ -60,12 +60,12 @@ std::string to_string(unsigned n)
   return s;
 }
 /*--------------------------------------------------------------------------*/
-string to_string(size_t n)
-{
-  char s[100];
-  sprintf(s, "%d", (int)n);
-  return s;
-}
+ string to_string(unsigned long int n)
+ {
+   char s[100];
+   sprintf(s, "%d", (int)n);
+   return s;
+ }
 /*--------------------------------------------------------------------------*/
 string to_string(int n)
 {
